jme3-core: remove unnecessary casts

+++ jme3-core/src/main/java/com/jme3/environment/generation/RunnableWithProgress.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009-2019 jMonkeyEngine
+ * Copyright (c) 2009-2020 jMonkeyEngine
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -66,7 +66,7 @@ public abstract class RunnableWithProgress implements Runnable {
      * @return fraction (&ge;0, &le;1)
      */
     public double getProgress() {
-        return (double) progress / (double) end;
+        return progress / (double) end;
     }
 
     /**
